writer

+++ src/zebra-shared.h
@@ -4,17 +4,25 @@
 
 typedef char byte;
 
+/**
+ * Int to bytes
+ */
 byte* int_bytes(int* x) {
   return (byte*)x;
 }
 
 
+/**
+ * Bytes to int
+ */
 int bytes_to_int(byte* bytes) {
   return bytes[0] 
-    + (bytes[1] << 8)
-    + (bytes[2] << 16)
-    + (bytes[3] << 24);
+    + (((int)bytes[1]) << 8)
+    + (((int)bytes[2]) << 16)
+    + (((int)bytes[3]) << 24);
 }
+
+
 
 #define zebra_keygen "shared_file.txt"
